+++ Core/RggScroll.pas
@@ -47,13 +47,13 @@ type
 
   TRggSB = class
   private
-    FMin: double;
-    FIst: double;
-    FMax: double;
+    FMin: single;
+    FIst: single;
+    FMax: single;
   public
-    Ist: double;
-    Min: double;
-    Max: double;
+    Ist: single;
+    Min: single;
+    Max: single;
     SmallStep: Integer;
     BigStep: Integer;
 
@@ -65,7 +65,7 @@ type
     procedure Assign(Value: TRggSB);
     procedure SaveToStream(s: TStream);
     procedure LoadFromStream(s: TStream);
-    function GetValue(n: TsbParam): double;
+    function GetValue(n: TsbParam): single;
 
     procedure Save;
     procedure Reset;
@@ -140,7 +140,7 @@ begin
   BigStep := 10;
 end;
 
-function TRggSB.GetValue(n: TsbParam): double;
+function TRggSB.GetValue(n: TsbParam): single;
 begin
   case n of
     TsbParam.Ist: result := Ist;
